@@ -9,14 +9,16 @@ use Illuminate\Routing\Controller;
 
 class AddressController extends Controller
 {
-    private $fillable_attributes = ["building", "street", "number", "postal_code", "city"];
+    private array $fillable_attributes = ["building", "street", "number", "postal_code", "city"];
 
-    public function getAddresses(): JsonResponse {
+    public function getAddresses(): JsonResponse
+    {
         $addresses = Address::all();
         return response()->json($addresses);
     }
 
-    public function getAddressById(int $id): JsonResponse {
+    public function getAddressById(int $id): JsonResponse
+    {
         $address = Address::find($id);
         if (!$address) {
             return response()->json(['message' =>'Address not found'], 404);
@@ -25,7 +27,8 @@ class AddressController extends Controller
         return response()->json($address);
     }
 
-    public function createAddress(Request $request): JsonResponse {
+    public function createAddress(Request $request): JsonResponse
+    {
         $address = new Address();
 
         foreach ($this->fillable_attributes as $attribute) {
@@ -36,7 +39,8 @@ class AddressController extends Controller
         return response()->json($address);
     }
 
-    public function updateAddress(Request $request, int $id) {
+    public function updateAddress(Request $request, int $id): JsonResponse
+    {
         $address = Address::find($id);
 
         if (!$address) {
@@ -53,7 +57,8 @@ class AddressController extends Controller
         return response()->json($address);
     }
 
-    public function deleteAddress($id) {
+    public function deleteAddress($id): JsonResponse
+    {
         $address = Address::find($id);
         if (!$address) {
             return response()->json(['message' => 'Address not found'], 404);
