@@ -9,14 +9,16 @@ use Illuminate\Routing\Controller;
 
 class OrganizerController extends Controller
 {
-    private $fillable_attributes = ["first_name", "last_name", "image", "phone", "email"];
+    private array $fillable_attributes = ["first_name", "last_name", "image", "phone", "email"];
 
-    public function getOrganizers(): JsonResponse {
+    public function getOrganizers(): JsonResponse
+    {
         $organizers = Organizer::all();
         return response()->json($organizers);
     }
 
-    public function getOrganizerById(int $id): JsonResponse {
+    public function getOrganizerById(int $id): JsonResponse
+    {
         $organizer = Organizer::find($id);
         if (!$organizer) {
             return response()->json(['message' =>'Organizer not found'], 404);
@@ -25,7 +27,8 @@ class OrganizerController extends Controller
         return response()->json($organizer);
     }
 
-    public function createOrganizer(Request $request): JsonResponse {
+    public function createOrganizer(Request $request): JsonResponse
+    {
         $organizer = new Organizer();
 
         foreach ($this->fillable_attributes as $attribute) {
@@ -36,7 +39,8 @@ class OrganizerController extends Controller
         return response()->json($organizer);
     }
 
-    public function updateOrganizer(Request $request, int $id) {
+    public function updateOrganizer(Request $request, int $id): JsonResponse
+    {
         $organizer = Organizer::find($id);
 
         if (!$organizer) {
@@ -53,7 +57,8 @@ class OrganizerController extends Controller
         return response()->json($organizer);
     }
 
-    public function deleteOrganizer($id) {
+    public function deleteOrganizer($id): JsonResponse
+    {
         $organizer = Organizer::find($id);
         if (!$organizer) {
             return response()->json(['message' => 'Organizer not found'], 404);
