@@ -2,20 +2,46 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Conference;
 use App\Models\Speaker;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Routing\Controller;
 use Illuminate\Http\Request;
 
 class SpeakerController extends Controller
 {
-    private $fillableAttributes = ['first_name', 'last_name', 'short_description', 'long_description', 'picture', 'linkedin', 'partner_id'];
+    private array $fillableAttributes = ['first_name', 'last_name', 'short_description', 'long_description', 'picture', 'linkedin', 'partner_id'];
 
-    public function getSpeakers() {
-        $speakers  = Speaker::with('partner')->get();
+    public function getSpeakers(): JsonResponse
+    {
+        $newestConference = Conference::orderBy('start_date', 'desc')->first();
+        $speakers = Speaker::with('partner')
+            ->whereHas('talk.timeSlots.stage.conferences', function ($query) use ($newestConference) {
+                $query->where('id', $newestConference->id);
+            })->get();
+
         return response()->json($speakers);
     }
 
-    public function createSpeaker(Request $request) {
+    public function getSpeakersByConference($conference_id): JsonResponse
+    {
+        $speakers = Speaker::whereHas('talk.timeSlots.stage.conferences', function ($query) use ($conference_id) {
+                $query->where('id', $conference_id);
+            })->get();
+        return response()->json($speakers);
+    }
+
+    public function getSpeakerById($id): JsonResponse
+    {
+        $speaker = Speaker::with('partner')->find($id);
+        if (!$speaker) {
+            return response()->json(['message' => 'Speaker not found'], 404);
+        }
+        return response()->json($speaker);
+    }
+
+    public function createSpeaker(Request $request): JsonResponse
+    {
         $speaker = new Speaker();
         foreach ($this->fillableAttributes as $attribute) {
             $speaker->$attribute = $request->input($attribute);
@@ -25,15 +51,7 @@ class SpeakerController extends Controller
         return response()->json($speaker);
     }
 
-    public function getSpeakerById($id) {
-        $speaker = Speaker::with('partner')->find($id);
-        if (!$speaker) {
-            return response()->json(['message' => 'Speaker not found'], 404);
-        }
-        return response()->json($speaker);
-    }
-
-    public function updateSpeaker(Request $request, $id)
+    public function updateSpeaker(Request $request, $id):JsonResponse
     {
         $speaker = Speaker::find($id);
 
@@ -52,7 +70,8 @@ class SpeakerController extends Controller
         return response()->json($speaker);
     }
 
-    public function deleteSpeaker($id) {
+    public function deleteSpeaker($id): JsonResponse
+    {
         $speaker = Speaker::find($id);
         if (!$speaker) {
             return response()->json(['message' => 'Speaker not found'], 404);
