@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Conference;
 use App\Models\Sponsor;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
@@ -9,14 +10,19 @@ use Illuminate\Routing\Controller;
 
 class SponsorController extends Controller
 {
-    private $fillable_attributes = ["name", "logo"];
+    private array $fillable_attributes = ["name", "logo"];
 
-    public function getSponsors(): JsonResponse {
-        $sponsors = Sponsor::all();
+    public function getSponsors(): JsonResponse
+    {
+        $newestConference = Conference::orderBy('start_date', 'desc')->first();
+        $sponsors = Sponsor::whereHas('conference', function ($query) use ($newestConference) {
+            $query->where('conference_id', $newestConference->id);
+        })->get();
         return response()->json($sponsors);
     }
 
-    public function getSponsorById(int $id): JsonResponse {
+    public function getSponsorById(int $id): JsonResponse
+    {
         $sponsor = Sponsor::find($id);
         if (!$sponsor) {
             return response()->json(['message' =>'Sponsor not found'], 404);
@@ -25,7 +31,8 @@ class SponsorController extends Controller
         return response()->json($sponsor);
     }
 
-    public function createSponsor(Request $request): JsonResponse {
+    public function createSponsor(Request $request): JsonResponse
+    {
         $sponsor = new Sponsor();
 
         foreach ($this->fillable_attributes as $attribute) {
@@ -36,7 +43,8 @@ class SponsorController extends Controller
         return response()->json($sponsor);
     }
 
-    public function updateSponsor(Request $request, int $id) {
+    public function updateSponsor(Request $request, int $id): JsonResponse
+    {
         $sponsor = Sponsor::find($id);
 
         if (!$sponsor) {
@@ -53,7 +61,8 @@ class SponsorController extends Controller
         return response()->json($sponsor);
     }
 
-    public function deleteSponsor($id) {
+    public function deleteSponsor($id): JsonResponse
+    {
         $sponsor = Sponsor::find($id);
         if (!$sponsor) {
             return response()->json(['message' => 'Sponsor not found'], 404);
