@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Conference;
 use App\Models\Stage;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
@@ -9,12 +10,18 @@ use Illuminate\Routing\Controller;
 
 class StageController extends Controller
 {
-    public function getStages(): JsonResponse {
-        $stages = Stage::all();
+    public function getStages(): JsonResponse
+    {
+        $newestConference = Conference::orderBy('start_date', 'desc')->first();
+        $stages = Stage::whereHas('conferences', function ($query) use ($newestConference) {
+            $query->where('conference_id', $newestConference->id);
+        })->get();
+
         return response()->json($stages);
     }
 
-    public function getStageById(int $id): JsonResponse {
+    public function getStageById(int $id): JsonResponse
+    {
         $stage = Stage::find($id);
         if (!$stage) {
             return response()->json(['message' =>'Stage not found'], 404);
@@ -23,7 +30,8 @@ class StageController extends Controller
         return response()->json($stage);
     }
 
-    public function createStage(Request $request): JsonResponse {
+    public function createStage(Request $request): JsonResponse
+    {
         $stage = new Stage();
 
         $stage->name = $request->input("name");
@@ -32,7 +40,8 @@ class StageController extends Controller
         return response()->json($stage);
     }
 
-    public function updateStage(Request $request, int $id) {
+    public function updateStage(Request $request, int $id):JsonResponse
+    {
         $stage = Stage::find($id);
 
         if (!$stage) {
@@ -47,7 +56,8 @@ class StageController extends Controller
         return response()->json($stage);
     }
 
-    public function deleteStage($id) {
+    public function deleteStage($id): JsonResponse
+    {
         $stage = Stage::find($id);
         if (!$stage) {
             return response()->json(['message' => 'Stage not found'], 404);
