@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Conference;
 use App\Models\User;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Routing\Controller;
 use Illuminate\Http\Request;
 
@@ -11,13 +13,16 @@ class UserController extends Controller
 {
     private $fillableAttributes = ['first_name', 'last_name', 'email', 'password', 'role', 'remember_token'];
 
-    public function getUsers()
+    public function getUsers($conference_id): JsonResponse
     {
-        $users = User::all();
+        $users = User::whereHas('registration.talk.timeSlots.stage.conferences', function ($query) use ($conference_id) {
+            $query->where('conference_id', $conference_id);
+        })->get();
+
         return response()->json($users);
     }
 
-    public function createUser(Request $request)
+    public function createUser(Request $request): JsonResponse
     {
         $user = new User();
         foreach ($this->fillableAttributes as $attribute) {
@@ -32,7 +37,7 @@ class UserController extends Controller
         return response()->json($user);
     }
 
-    public function getUserById($id)
+    public function getUserById($id): JsonResponse
     {
         $user = User::find($id);
 
@@ -43,7 +48,7 @@ class UserController extends Controller
         return response()->json($user);
     }
 
-    public function updateUser(Request $request, $id)
+    public function updateUser(Request $request, $id): JsonResponse
     {
         $user = User::find($id);
         if (!$user) {
@@ -65,7 +70,7 @@ class UserController extends Controller
         return response()->json($user);
     }
 
-    public function deleteUser($id)
+    public function deleteUser($id): JsonResponse
     {
         $user = User::find($id);
 
