@@ -30,17 +30,41 @@ class AddUsers extends Seeder
                 'remember_token' => null,
             ],
             [
-                'first_name' => 'Ján',
-                'last_name' => 'Skalka',
-                'email' => 'jskalka@ukf.sk',
+                'first_name' => 'Lukáš',
+                'last_name' => 'Sýkora',
+                'email' => 'sykora@ukf.sk',
                 'password' => null,
                 'role' => 'attendee',
                 'remember_token' => null,
             ],
             [
-                'first_name' => 'Milan',
-                'last_name' => 'Turčáni',
-                'email' => 'mturcani@ukf.sk',
+                'first_name' => 'Katarína',
+                'last_name' => 'Vargová',
+                'email' => 'vargova@ukf.sk',
+                'password' => null,
+                'role' => 'attendee',
+                'remember_token' => null,
+            ],
+            [
+                'first_name' => 'Peter',
+                'last_name' => 'Horváth',
+                'email' => 'horvath@ukf.sk',
+                'password' => null,
+                'role' => 'attendee',
+                'remember_token' => null,
+            ],
+            [
+                'first_name' => 'Jana',
+                'last_name' => 'Kováčová',
+                'email' => 'kovacova@ukf.sk',
+                'password' => null,
+                'role' => 'attendee',
+                'remember_token' => null,
+            ],
+            [
+                'first_name' => 'Martin',
+                'last_name' => 'Novák',
+                'email' => 'novak@ukf.sk',
                 'password' => null,
                 'role' => 'attendee',
                 'remember_token' => null,
